refactor the route structure for more clarity among routines and workouts
